feat: add updateArtist; fix: fix some error in artist controller

+++ src/controllers/artist.controller.ts
@@ -8,6 +8,7 @@ import {
   NotFoundException,
   Param,
   Post,
+  Put,
 } from '@nestjs/common';
 import { isUUID } from 'class-validator';
 import { ArtistService } from 'src/services/artist.service';
@@ -37,8 +38,9 @@ export class ArtistController {
 
   @Post()
   addArtist(@Body() artistData: CreateArtistDto) {
-    const isValidData = artistData.name || artistData.grammy;
-    if (!isValidData)
+    const isValidData =
+      typeof artistData.grammy !== 'boolean' || !artistData.name;
+    if (isValidData)
       throw new BadRequestException('Required fields not filled');
     return this.artistService.addNewArtist(artistData);
   }
@@ -46,11 +48,28 @@ export class ArtistController {
   @Delete(':id')
   @HttpCode(204)
   deleteArtist(@Param('id') id: string) {
+    if (!isUUID(id, 4)) throw new BadRequestException('Invalid arist id');
     const artist = this.artistService.getArtist(id);
     if (!artist)
       throw new NotFoundException(`Artist with id - ${id} doesn't exist`);
 
     this.artistService.deleteArtist(id);
     return;
+  }
+
+  @Put(':id')
+  updateArtist(@Param('id') id: string, @Body() artistData: CreateArtistDto) {
+    if (!isUUID(id, 4)) throw new BadRequestException('Invalid artist id');
+
+    if (!artistData.name || typeof artistData.grammy !== 'boolean') {
+      throw new BadRequestException('Fields not valid');
+    }
+
+    const newArtist = this.artistService.updateArtist(artistData, id);
+
+    if (!newArtist)
+      throw new NotFoundException(`Artist with id - ${id} not found`);
+
+    return newArtist;
   }
 }
